Add files via upload
Today's work included work on the clean data files, creation of an overall runner file, making a flow diagram for the program, research into CPA, reworking FindDistribution to lean on Scipy instead of on Fitter, and combining the pullData and writetoTXT function. 

The most important part of today's work was definitely the flow diagram and CPA research. I now have a much cleaner picture of the program in my head and a better understanding of CPA (albeit quite a few more questions). Monday I plan on shifting the data format to a pandas dataframe in addition to getting connected to the actual OMOP database. If I have extra time, I'll either put more work into the cleanData file or write pseudocode for CPA.