Workday 6-5-17
Today mostly consisted of work on the data cleaning file and the conversion of the code to dataframes. Most of the functionality within the clean data file is working. Tomorrow should consist of hopefully finishing the categorical to continuous function. I have most of the concept down for it, I just need to work on implementation. After that is CPA!!!! Also I still am not officially connected to OMOP. Will talk to Richard this week. Will update again tomorrow