Add files via upload
Today's focus was on the categorical to continuous function in the CleanData file. There is a lot that goes into the thought process behind converting categorical data to continuous data and I'm still somewhat aprehensive about the process SDV chose to employ. It could lead to massively large extended tables when run through CPA and I think it could lead to some pretty severe overfitting. The issue is going to be escalated to Jason tomorrow mornign and I guess we'll just have to see what he says. 

Tomorrow should hopefully be finishing the CleanData file and moving on to CPA. Hopefully.