Workday 6-7-17
The majority of today was focused on the CPA script. I am debugging it on perfectly clean data. So far it isn't too terrible. I still don't have an answer on the categorical variables issue, but I am connected to OHDSI. Tomorrow should be more work on the find distributions function, CPA, and categorical variables