Work 6-8-17
The majority of work today focused on CPA. I was able to get the algorithm running on fake (perfectly clean) continuous data. I was also able to start working on the categorical side of the algorithm but still have a long way to go. I also talked to Jason this morning about the continuous hangup and he basically said to power through it. I don't know exactly how this will work, but I definitely need to make the code more efficient. Things for a later day I guess.

Tomorrow I plan to focus entirely on the categorical side of CPA. Having looked at the OMOP model, that will probably be the most important part of the entire algorithm. Almost every single feature is categorical