Work Day 6/9 :)
Today was focused on categorical CPA. although that side of the algorithm isn't quite finished yet, It did advance quite a bit. The unique values are found, counted, and logged. They just aren't logged in the exact way I want them to be. The would be a small problem for Monday. 

On Monday, I need to figure out how to format the data, and then start working on how the metadata is transferred. All of the possible categories for each column in each table need to be logged and stored efficiently. I'm thinking this might be possible without a separate metadata file by just using the column headers. then If I have time I'll start working how to combine the categorical and continuous data in unison. shouldn't theoretically be too much of a problem, but for some reason the concept just unnerves me. We'll see