Work 6-13-17
Today I finished the CPA algorithm. I got to work early this morning and spent about an hour creating an exact workflow of what needed to happen in the CPA algorithm. I reworked the order of how I wanted the metadata to be created and stored with relation to the data calculations. That time spent in organization payed off ten fold. I now have a more efficient and cleaner version of the file uploaded. That was a huge step of the overall algorithm that I didn't expect to have finished until the end of this week, so I'm doing well! 

I would like to spend tomorrow cleaning and standardizing the code I've already written to make sure it is easily readable, modular and efficient. Then after that I will go back to working on the overall runner file. Making sure that the general workflow is in proper order, and putting together the first half of the overall algorithm! 

Today was a big day!