Work 6-15-17
The past two days has been a mishmash of different things. I've played around with the OMOP-CDM, tried to make the existing code error out in a number of ways, cleaned miscellaneous things within the code, researched object oriented programming and recursion, and planned how I want to proceed with the actual program. Things are starting to pile up with regards to organization and workflow of the code, so it's starting to be more than I feel like I can handle.

I feel like I have most, if not all of the pieces needed for the modeling already made. There are just so many different ways the functions can error out or so many little things that can go wrong #frustration. I think tomorrow needs to be another strictly organizational day. I want the exact outputs for the modeling phase (and inputs for the synthesis phase) of SDV to be clearly defined after tomorrow. Then I want to figure out how to use the pieces of code I've made (in addition to maybe identifying a few more to make) in order to make