Work 6-16-17
The first half of today was the IRAD extravaganza. The second half I reread the SDV paper with an emphasis on the order of the functions. I make somewhat of a large realization. I don't need to save every single file locally. In fact, the only things I should be saving locally are the extended tables. that means I'm better off reworking the entire layout of the code. CPA should come before data cleaning. That way I don't have to clean every single table every time I pull form the database. Instead, I can get the data with nulls, date times and everything in between, combine it together and then clean it once. this should make the code as a whole run much faster.

I also talked with Jon Duke this afternoon and he suggested that I focus all of my time and efforts on categorical variables (effectively ignoring continuous ones) because that is the majority of an OMOP dataset. I'll come in on Monday and start working on that. Also the specific variables to pass from modeling to synthesis and distri