Work 6-21-17
Today marked the start of the actual testing on the algorithm as a whole. I had to make white a few changes (especially to the scheme code), but as a whole it worked pretty well! Unfortunately, this is a huge algorithm and my Mac is having problems actually completing it. I plan on trying to run it overnight tonight just to see if it'll complete. If not, then I definitely need to make some changes to the CPA side of the algorithm. Also maybe something about date time to EPOCH can be slow too. I've noticed that as well. 

Tomorrow I'm not entirely sure what's going to happen. What I expect is to get an incomplete or imperfect set of CSVs and have to determine what went wrong in the CPA algorithm. If I can do that before Top golf, then I might be able to actually get results from Jacob's computer on Friday. Who knows. For now I just need to run a program and see.